combine slatepreviewblocks & make internal/external styling more consistent
